Delay initial boot time calc to after setup
Calculating the initial boot time on device first turning on is not
meaningfully useful as manufacturers may turn on the device and modify
the real time clock as part of IMEI writing / other processes. This may
technically constitute a "first boot" but it won't get to the hands of
the users until much later.

Instead, what we really want is the first time the device is actually
being used by the user, and post set-up seems like a good signal for
this.

Due to the process dying before set-up completes and the lack of a
platform mechanism to wake up the app on set-up finishing, the initial
unused tracking time may actually end up occurring the first time the
hibernation job runs.

Bug: 243888248
Flag: EXEMPT bugfix
Test: atest HibernationPolicyTest
Test: atest CtsHibernationTestCases
Test: manual (verified no value in shared preferences before setup and
correct value in shared preferences after)

Change-Id: Ibcb6a4cec901395400b60f423e9e7f7ced0ec212