Add drawable states to prefrence background for rounded corners
One of state_single, state_first, state_middle and state_last will be
set on the drawables of the preference root layout, so that the
background can be a selector that changes shape in order to implement
rounded corners.

Another new overlayable attribute app:state_has_icon_space has also been
added, so that the background can include a divider that changes its
start inset accordingly.

For the sake of reducing the number of changes to cherry-pick, this
change also includes a tiny fix for not styling
MultiTargetSwitchPreference since it's (normally) not used in
permission settings.

Bug: 327285817
Flag: NONE infeasible
Test: manually with ag/28621492
Change-Id: Ib5c94587bd2c149ce4b1075c12cd31db8f3c913d