Fix clustering logic for permission timeline page
Consider duration when clustering accesses i.e. if
camera was accessed for 5 minutes at minute 10, then
minute 15 or 16 will be combined with this cluster.

When calculating duartion for camera/mic, use first and last
entry and add the duration for last reported event i.e. if
access happened at minute 2 for 3 minutes and at minute 5 for
10 minutes then the duration would be 5 - 2 + 10 = 13.
This is to avoid double count in the duration, in this example
if app op events are reported at minute 3 and 4 for 1 minute each
then summing up all the duration (3 + 1 + 1 + 10) would result
in 15 minute.

Also, don't pass duplicate tags in ACTION_MANAGE_PERMISSION_USAGE
intent.

Fix: 355526988
Fix: 358502156
Test: manual
FLAG: EXEMPT bugfix
Change-Id: Ieb7e03f0bd0a8924b1cee758530301405ecaa42e