Change the content provider for test running user
Currently content provider is used for SYSTEM User. This doesn't work if the test apk is not installed for System User. There is no need for content provider to run as system User, so running it as the test running User.

Flag: EXEMPT test fix

LOW_COVERAGE_REASON=TEST_ONLY

Bug: 359670220
Bug: 355564081
Test: atest CtsPermissionUiTestCases:android.permissionui.cts.NotificationPermissionTest
Test: atest --user-type secondary_user_on_secondary_display CtsPermissionUiTestCases:android.permissionui.cts.NotificationPermissionTest
Change-Id: I26d26a7b7cd4f689db241b667b3851e44ef65349