Stop using DeviceConfig to enable / disable SafetyCenter on UDC+.
Rely solely on the XML configuration instead.

This flag is true on UDC+ on 100% of devices, so this change should technically be a no-op.

This is done for a few reasons:
- Don’t run through the DeviceConfig code path when it’s not necessary
- Avoid potential regressions where we wrongly turn this flag off server-side
- Avoid issues such as b/341315099 where the flag ends up being false server-side for some reason

Bug: 354670125
Test: atest CtsSafetyCenterTestCases
Relnote: Stop using DeviceConfig to enable / disable SafetyCenter on UDC+
Flag: com.android.permission.flags.safety_center_enabled_no_device_config
LOW_COVERAGE_REASON=b/360423897

Change-Id: I56a1e2927cf3f33d24f2f249bbecfbeeffe04e2a