Add workaround for new permission without flag
For permissions introduced after 24Q3 a hack is added to set their flag
as "missing" and in PermissionPolicyTest we assume the permission is
enabled since flagless permissions are visible. This allows the lint
check and PermissionPolicyTest to pass so that the lint check change
can pass presubmit checks. Since PermissionPolicyTest continues to
assume the permission is visible it will remain release blocking on 24Q3.

There is also logic in the test to determine when this workaround
isn't needed anymore so that it can be known to be removed in the same
change.

LOW_COVERAGE_REASON=TEST_ONLY
Relnote: Test only, no functional changes
Test: Run lint check
      PermissionPolicyTest on ToT and on 24Q3,
      PermissionPolicyTest when replacing "missing" with a real flag in
          android_manifest.xml
Flag: TEST_ONLY
Bug: 361567988
Bug: 333749198
Change-Id: Iacdbeca507cb29c59ec8121762219c08da99cf7a