[BC25] Split (Legacy)AppPermissionFragment
This copies the AppPermissionFragment class to
LegacyAppPermissionFragment, and updates our codebase to
conditionally use one class or the other in the following way:

- LegacyAppPermissionFragment will be used for max SDK 35
- AppPermissionFragment will be used for min SDK 36

The motivation behind this is that we wish to rewrite
AppPermissionFragment to use Preferences, but doing so is not a
backwards-compatible change, particularly when it comes to RROs
(style/resource overlays).

In most cases, we still probably want code in these fragments to stay
mostly in sync. To remind us of this, both files are wrapped with
"LINT.IfChange", which will require engineers who modify one file to
also modify the other. (This can be overridden with "NO_IFFT: <reason>"
in the git commit message.)

Bug: 349675008
Test: atest CtsPermissionUiTestCases
Flag: com.android.permission.flags.app_permission_fragment_uses_preferences
Relnote: N/A
LOW_COVERAGE_REASON=b/365859967
Change-Id: I217f041308d78d6ee30663f33f66f799bb257398