PermissionReviewTapjackingTest: replace pressHome() with pressBack().
This change ensures that the top activity (".FinishOnCreateActivity") properly goes away on all devices. On some devices, pressHome() does not hide the visibility of the apps.

The test checks that the overlay "Find me!" is present, then opens up a top activity that should hide the overlay. The test asserts that the overlay is no longer present, and finally, hides the top activity, and asserts that the overlay is present again.

Relnote: Fix CTS test `PermissionReviewTapjackingTest` to close out of apps properly.

Bug: 350551451
Test: atest PermissionReviewTapjackingTest on Android 14 devices
Flag: EXEMPT bugfix
Change-Id: I3cf77cd69466b0bfc2bee8e479c05ced62befafb